ahtak/requirements.txt
Django>=5.2,<6
psycopg2-binary>=2.9
python-dotenv>=1.0
djangorestframework>=3.14
django-filter>=24.0
django-cors-headers>=4.3
Pillow>=10.0
celery>=5.3
redis>=5.0
reportlab>=4.0
openpyxl>=3.1
qrcode[pil]>=7.4
django-celery-beat>=2.5
python-dateutil>=2.8
django-jazzmin>=3.0
